+++ src/Plugin.php
@@ -7,12 +7,17 @@
 
 namespace craft\commerce\stripe;
 
+use craft\commerce\events\UpdatePrimaryPaymentSourceEvent;
+use craft\commerce\Plugin as CommercePlugin;
+use craft\commerce\services\Customers;
 use craft\commerce\services\Gateways;
+use craft\commerce\stripe\base\Gateway;
 use craft\commerce\stripe\gateways\PaymentIntents;
 use craft\commerce\stripe\models\Settings;
 use craft\commerce\stripe\plugin\Services;
 use craft\events\RegisterComponentTypesEvent;
 use yii\base\Event;
+use yii\base\InvalidConfigException;
 
 /**
  * Plugin represents the Stripe integration plugin.
@@ -48,6 +53,31 @@ class Plugin extends \craft\base\Plugin
                 $event->types[] = PaymentIntents::class;
             }
         );
+
+        Event::on(
+            Customers::class,
+            Customers::EVENT_UPDATE_PRIMARY_PAYMENT_SOURCE,
+            [$this, 'handlePrimaryPaymentSourceUpdated']
+        );
+    }
+
+    /**
+     * Whenever a payment source is set as primary in Commerce, lets make it the primary in the gateway too.
+     *
+     * @param UpdatePrimaryPaymentSourceEvent $event
+     * @return void
+     * @throws InvalidConfigException
+     */
+    public function handlePrimaryPaymentSourceUpdated(UpdatePrimaryPaymentSourceEvent $event): void
+    {
+        $paymentSourceService = CommercePlugin::getInstance()->getPaymentSources();
+        $newPrimaryPaymentSource = $paymentSourceService->getPaymentSourceById($event->newPrimaryPaymentSourceId);
+        /** @var Gateway $gateway **/
+        $gateway = $newPrimaryPaymentSource->getGateway();
+        if ($gateway instanceof Gateway) {
+            $stripeCustomerReference = $this->getCustomers()->getCustomer($gateway->id, $event->customer)->reference;
+            $gateway->setPaymentSourceAsDefault($stripeCustomerReference, $newPrimaryPaymentSource->token);
+        }
     }
 
     /**