+++ src/services/Customers.php
@@ -12,6 +12,7 @@ use craft\commerce\Plugin as Commerce;
 use craft\commerce\stripe\base\Gateway as BaseGateway;
 use craft\commerce\stripe\base\SubscriptionGateway;
 use craft\commerce\stripe\errors\CustomerException;
+use craft\commerce\stripe\events\CreateCustomerEvent;
 use craft\commerce\stripe\models\Customer;
 use craft\commerce\stripe\Plugin as StripePlugin;
 use craft\commerce\stripe\records\Customer as CustomerRecord;
@@ -114,6 +115,26 @@ class Customers extends Component
     {
         $customerRow = $this->_createCustomerQuery()
             ->where(['id' => $id])
+            ->one();
+
+        if ($customerRow) {
+            return new Customer($customerRow);
+        }
+
+        return null;
+    }
+
+    /**
+     * Return a customer by its user ID.
+     *
+     * @param int $id
+     *
+     * @return Customer|null
+     */
+    public function getCustomerByUserId(int $id): ?Customer
+    {
+        $customerRow = $this->_createCustomerQuery()
+            ->where(['userId' => $id])
             ->one();
 
         if ($customerRow) {